+++ SmartDigitalPsico.Domain/ModelEntity/MedicalCalendar.cs
@@ -25,7 +25,7 @@ namespace SmartDigitalPsico.Domain.ModelEntity
         public DayOfWeek[] RecurrenceDays { get; set; } = []; // 0 = Sunday  1 = Monday  6 = Saturday
         public ERecurrenceCalendarType RecurrenceType { get; set; }
         public DateTime? RecurrenceEndDate { get; set; }
-        public int? RecurrenceCount { get; set; }
+        public byte? RecurrenceCount { get; set; }
 
         #endregion Columns
